# backend-app/docs/swagger.yaml
basePath: /api/v1
definitions:
  internal_modules_user.CreateUserRequest:
    properties:
      email:
        example: john@example.com
        type: string
      name:
        example: John Doe
        maxLength: 100
        minLength: 2
        type: string
      password:
        example: secret123
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  internal_modules_user.ErrorResponse:
    properties:
      details:
        example: additional details
        type: string
      error:
        example: error message
        type: string
    type: object
  internal_modules_user.LoginRequest:
    properties:
      email:
        example: john@example.com
        type: string
      password:
        example: secret123
        type: string
    required:
    - email
    - password
    type: object
  internal_modules_user.LoginResponse:
    properties:
      token:
        example: eyJhbGciOiJIUzI1NiIs...
        type: string
      user:
        $ref: '#/definitions/internal_modules_user.UserResponse'
    type: object
  internal_modules_user.PaginatedResponse:
    properties:
      data: {}
      limit:
        example: 10
        type: integer
      page:
        example: 1
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
    type: object
  internal_modules_user.SuccessResponse:
    properties:
      message:
        example: operation successful
        type: string
    type: object
  internal_modules_user.UpdateUserRequest:
    properties:
      email:
        example: john.updated@example.com
        type: string
      name:
        example: John Updated
        maxLength: 100
        minLength: 2
        type: string
    type: object
  internal_modules_user.UserResponse:
    properties:
      created_at:
        example: "2024-01-01T00:00:00Z"
        type: string
      email:
        example: john@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      updated_at:
        example: "2024-01-01T00:00:00Z"
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@example.com
    name: API Support
    url: http://www.example.com/support
  description: A backend API server built with Go and Gin
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Go Backend API
  version: "1.0"
paths:
  /users:
    get:
      consumes:
      - application/json
      description: Retrieve all users with pagination
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_user.PaginatedResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
      summary: Get all users
      tags:
      - users
  /users/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a user by ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_user.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
      summary: Delete user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Retrieve a user by their ID
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update an existing user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: User update data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
      summary: Update user
      tags:
      - users
  /users/login:
    post:
      consumes:
      - application/json
      description: Authenticate a user and return a token
      parameters:
      - description: User login credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_modules_user.LoginResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
      summary: User login
      tags:
      - users
  /users/register:
    post:
      consumes:
      - application/json
      description: Create a new user account
      parameters:
      - description: User registration data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/internal_modules_user.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_modules_user.ErrorResponse'
      summary: Register a new user
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"